Move the searchForFacetValues struct into a dedicated module
